src/server/requirements.txt
# --- Core Framework & Web Server ---
fastapi
uvicorn[standard]

# --- Data Validation & Config ---
python-dotenv

# --- Authentication & Security ---
python-jose[cryptography]
cryptography

# --- Asynchronous & HTTP Clients ---
httpx
requests
requests-oauthlib

# --- Database & Caching ---
motor
redis
asyncpg
pgvector

# --- AI & Machine Learning ---
qwen-agent
google-genai
elevenlabs
librosa
torch
fastrtc[vad, stt, tts]
numpy
crawl4ai
playwright

# --- Background Task Queue ---
celery
python-dateutil # For rrule in scheduler

# --- Google API Clients ---
google-api-python-client
google-auth-oauthlib

# --- Other 3rd Party API Clients ---
PyGithub
notion-client

# --- Analytics ---
posthog

# --- MCP Framework ---
fastmcp

# --- Charting (for gslides MCP) ---
matplotlib

jsonextractor

# --- Vector DB and Embeddings ---
chromadb-client
sentence-transformers
google-generativeai
rich

textract==1.5.0